Build and run scripts for a little less typing.
For some reason the Swift Package Manager is lacking the means to check for
updated packages, and I'm developing the parallel Stockfighter library all the
time, so I needed a way to force it to reload… well, there you go. Deleting the
`Packages` dir and redownloading packages seems like the best way for now.